Some bug fixes, optimizations, and switch all integers to use size_t

+++ src/NewsItem.h
@@ -12,7 +12,7 @@
 struct NewsItem {
   std::string path; //! holds the path to the information, whether a file path or web path
   std::string collection; //! The classification of the news item set as empty or unknown when unknown
-  std::map<std::string, unsigned> word_count; //! The count (per word) of the contents of the news item
+  std::map<std::string, size_t> word_count; //! The count (per word) of the contents of the news item
   std::string contents; //! The full contents of the news item
 };
